requirements.txt
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.17.0
pandas==0.25.0
pdf2image==1.6.0
Pillow==6.1.0
pypandoc==1.4
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
Werkzeug==0.15.5
